requirements.txt
python==<Python 3.9.16>
altair==5.0.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
catalogue==2.0.8
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
executing==1.2.0
fastcore==1.5.29
filelock==3.12.0
fonttools==4.39.4
fsspec==2023.5.0
fuzzywuzzy==0.18.0
h5py==3.8.0
huggingface-hub==0.14.1
identify==2.5.24
idna==3.4
imageio==2.28.1
imgaug==0.4.0
importlib-resources==5.12.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonlines==3.1.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
Levenshtein==0.21.0
lit==16.0.5
loguru==0.7.0
lovely-numpy==0.2.9
lovely-tensors==0.1.15
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
pre-commit==3.3.2
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.15.1
PyMuPDF==1.22.3
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-Levenshtein==0.21.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==3.0.0
regex==2023.5.5
requests==2.30.0
rich==13.3.5
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
shapely==2.0.1
six==1.16.0
srsly==2.4.6
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
toolz==0.12.0
torch==2.0.1
torch-snippets==0.499.27
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
wasabi==1.1.1
wcwidth==0.2.6
xmltodict==0.13.0
zipp==3.15.0
